worker/requirements.txt
asyncpg>=0.30.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0,<5.0.0
